telegram_bot/requirements.txt
# telegram-bot/requirements.txt

python-telegram-bot[job-queue]==22.3
aiohttp>=3.9
requests>=2.31.0
psycopg2-binary>=2.9
Babel>=2.17,<3
python-json-logger>=2.0.7
prometheus-client>=0.20.0
sentry-sdk>=2.13.0
tldextract>=3.6.0
